Big changes - Shader Graph
Scatter lighting attempt in Shader Graph
Change from Shader Graph to Visual Shader for visual effect on plane (Except for Scatter lighting)
Working Compute Shader
Working hight readabilty and adjustment
Working Bouyency for non-controlable objects
Lacklaster Bouyency for moving objects

Haven't pushed for a while. In creating the Bouyency feature I got stuck on the physics solution for to long and desided to make my own solution. It doesn't work on moving objects, since it is not adaptable to the async input from the compute shader.

Gonna work on the Light Scattering now